requirements.txt
streamlit
qdrant-client
beautifulsoup4
requests
python-dotenv
langchain
langchain-qdrant
langchain-google-genai
langchain-community